fix: resolve module-evaluation-order bugs the compiler swap exposed
Not tooling config -- real fragility in application/test code that
TSC's specific CommonJS emit style happened to paper over, surfaced by
switching to SWC/esbuild for these files. Each is a genuine gap, not a
compiler quirk to work around:

- Circular import: RedirectService.ts imported CookieArrayReducerService
  and CrawlerCheckService from the same services/index.ts barrel that
  re-exports RedirectService itself, and separately imported
  RedirectType/RedirectResult from the redirect/index.ts barrel that
  imports RedirectController which imports RedirectService. TSC's
  lazy-getter re-export codegen tolerates the resulting cycle; SWC's
  compiled output (also getter-based, but using `const` instead of
  `var` for the require binding) hits a genuine TDZ and throws
  "Cannot access before initialization". Fixed by extracting
  RedirectType/RedirectResult into a new redirect/types.ts, and
  importing CookieArrayReducerService/CrawlerCheckService directly
  from their own files instead of through the sibling barrel -- both
  are the standard fix for barrel-induced import cycles, independent
  of which compiler is involved.

- Non-configurable named exports: AnalyticsLoggerService.test.ts used
  `jest.spyOn(analytics, 'generateCookie')` on a namespace import.
  SWC compiles named exports as non-configurable property getters
  (closer to real ECMAScript module semantics than TSC's more
  permissive `exports.foo = foo`), so spyOn can never redefine it.
  Rewrote using jest.mock + jest.requireActual to wrap the real
  implementation in a trackable jest.fn -- the correct pattern
  regardless of compiler, since real ESM namespaces are non-writable
  too.

- Import-hoisting order: src/server/index.ts (and 3 test setup files
  with the identical pattern) called bindInversifyDependencies()
  textually before `import api from './api'`, relying on TSC not
  hoisting that import above the call. api/index.ts resolves
  inversify bindings at module-load time, so it needs binding
  registration to have already run. SWC (and, more fundamentally,
  real ESM) hoists all imports above other statements regardless of
  source position, so `./api` was loading before the bindings existed
  -- "No matching bindings found" / "Ambiguous match found" depending
  on which container mutation raced which. Fixed by making the
  ordering explicit with a dynamic require() instead of a static
  import in each of the 4 affected files, which cannot be hoisted by
  any compiler.

Verified on typescript@5.9.3 (still, ahead of the version bump): full
jest suite (57 suites / 477 tests), build, lint, format:check, and a
manual boot of both `tsx watch`-run and built server entry points
through the real DI chain.
