+++ src/server/modules/redirect/RedirectController.ts
@@ -6,7 +6,7 @@ import { NotFoundError } from '../../util/error'
 import parseDomain from '../../util/domain'
 import { DependencyIds, ERROR_404_PATH } from '../../constants'
 import { AnalyticsLoggerService, RedirectService } from './services'
-import { RedirectType } from '.'
+import { RedirectType } from './types'
 import { EventAction, EventCategory } from './ga/types/enum'
 import { createPageViewHit } from './ga'
 import assetVariant from '../../../shared/util/asset-variant'
