+++ src/server/modules/redirect/services/RedirectService.ts
@@ -2,10 +2,11 @@ import { inject, injectable } from 'inversify'
 import { UrlRepositoryInterface } from '../../../repositories/interfaces/UrlRepositoryInterface'
 import { DependencyIds } from '../../../constants'
 import { NotFoundError } from '../../../util/error'
-import { RedirectResult, RedirectType } from '..'
+import { RedirectResult, RedirectType } from '../types'
 import { LinkStatisticsService } from '../../analytics/interfaces'
 import { logger, ogUrl } from '../../../config'
-import { CookieArrayReducerService, CrawlerCheckService } from '.'
+import { CookieArrayReducerService } from './CookieArrayReducerService'
+import { CrawlerCheckService } from './CrawlerCheckService'
 import { UrlThreatScanService } from '../../threat/interfaces'
 import { getSafeBrowsingExpiryDate } from '../../../util/safeBrowsing'
 import { UrlManagementService } from '../../user/interfaces'
