+++ src/server/modules/redirect/types.ts
@@ -1,0 +1,10 @@
+export enum RedirectType {
+  Direct,
+  TransitionPage,
+}
+
+export type RedirectResult = {
+  visitedUrls: string[]
+  longUrl: string
+  redirectType: RedirectType
+}
